Fixed Class and Constructor to include "_Lite"
I write the library and changed the name to append "_Lite" to differentiate between the existing "TLC5916" library.  I decided to go ahead and submit my library after seeing the existing library because I wanted to more closely match the SOLID principles, and anything that wasn't specific to communicating bytes of into with the iC was left out.